+++ src/Messaging.Abstractions/IEnvelopePublicationPipeline.cs
@@ -1,0 +1,8 @@
+using ArchiScrapper.Models;
+
+namespace ArchiScrapper.Messaging.Abstractions;
+
+public interface IEnvelopePublicationPipeline
+{
+    RawEnvelope Compose<TPayload>(TypedEnvelope<TPayload> source, Func<TPayload, string> payloadSerializer);
+}